+++ CLI.txt
@@ -33,7 +33,7 @@ CODE
 ============================================================
 
 poetry build
-pip install --force-reinstall dist/snakemake_report_plugin_metadata4ing-1.2.5-py3-none-any.whl
+pip install --force-reinstall dist/snakemake_report_plugin_metadata4ing-2.0.0-py3-none-any.whl
 snakemake --reporter metadata4ing --cores 1 ./paper.pdf
 unzip ro-crate-metadata.zip -d ro-crate-metadata
