+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "snakemake-report-plugin-metadata4ing"
-version = "1.2.6"
+version = "2.0.0"
 description = "Snakemake reporter plugin to extract metadata accoriding to MetaData4Ing ontology"
 readme = "README.md"
 requires-python = ">=3.11,<4.0"
