+++ src/snakemake_report_plugin_metadata4ing/utils.py
@@ -1,0 +1,34 @@
+import mimetypes
+import os
+import re
+from pathlib import Path
+
+
+def get_mime_type(file_name: str) -> str:
+    file_name = Path(file_name).name
+    mime_type, _ = mimetypes.guess_type(file_name, strict=False)
+    return mime_type or "application/octet-stream"
+
+
+def validate_filename(filename: str) -> None:
+    if not filename or filename.strip() == "":
+        raise ValueError("Filename cannot be empty.")
+
+    illegal_pattern = r'[<>:"/\\|?*]'
+    if re.search(illegal_pattern, filename):
+        raise ValueError(f"Filename '{filename}' contains illegal characters.")
+
+    reserved_names = {
+        "CON",
+        "PRN",
+        "AUX",
+        "NUL",
+        *{f"COM{i}" for i in range(1, 10)},
+        *{f"LPT{i}" for i in range(1, 10)},
+    }
+
+    if filename.upper().split(".")[0] in reserved_names:
+        raise ValueError(f"Filename '{filename}' is reserved on Windows.")
+
+    if os.path.isdir(filename):
+        raise ValueError(f"'{filename}' is a directory, not a file.")
